+++ packages/entity_store_firestore/lib/src/firestore_repository/transaction.dart
@@ -1,5 +1,12 @@
 part of '../firestore_repository.dart';
 
+final class FirestoreTransactionOptions extends TransactionOptions {
+  final Duration? timeout;
+  final int? maxAttempts;
+
+  FirestoreTransactionOptions({this.timeout, this.maxAttempts});
+}
+
 class FirestoreTransactionContext extends ITransactionContext {
   final Transaction value;
 
@@ -23,7 +30,15 @@ class FirestoreTransaction extends ITransaction<FirestoreTransactionContext> {
   Future<Result<(T, FirestoreTransactionContext), Exception>>
       handleTransaction<T>(
     Future<T> Function(FirestoreTransactionContext context) fn,
+    TransactionOptions? options,
   ) async {
+    FirestoreTransactionOptions? firestoreOptions;
+    if (options != null && options is! FirestoreTransactionOptions) {
+      throw ArgumentError("options must be FirestoreTransactionOptions");
+    } else if (options != null) {
+      firestoreOptions = options as FirestoreTransactionOptions;
+    }
+
     try {
       return await instance.runTransaction(
         (transaction) async {
@@ -31,6 +46,8 @@ class FirestoreTransaction extends ITransaction<FirestoreTransactionContext> {
           final result = await fn(context);
           return (result, context).toOk();
         },
+        timeout: firestoreOptions?.timeout ?? const Duration(seconds: 30),
+        maxAttempts: firestoreOptions?.maxAttempts ?? 5,
       );
     } catch (e) {
       if (e is Exception) {